Parse constants and bracket-less symbolics
I forgot to handle constants beyond the tokenizer. They are generally
not useful because they can always be eliminated which makes formulas
smaller, but maybe it is more convenient for automatically produced
formula strings to support them. They got a new syntax \T and \F.

That is because I decided that for compatibility with sage's propcalc,
I want to support symbolic variables without brackets, if they are
sufficiently alphanumeric. The following works now:

    infix: (ab&3 | a34&x) -> (\T ^ x)
    rpn:   [ab] 3 & [a34] [x] & | \T [x] ^ >

Symbolics are always printed using brackets though, just like operators
are always printed in their short form.