Introduce Variable and Domain
After much deliberation, I decided to ditch the "variables are just
numbers" strategy. This seemed appealing from a memory-compactness,
bit-twiddling-efficiency and DIMACS serialization point of view.

But thinking more thoroughly about the intended usage of this library,
variable numbers seem a useless limitation. E.g. for CNF conversion
based on truth tables, variables have to be renumbered into a consecutive
sequence 1..N anyway (to generate all assignments easily using a binary
incrementer).

This led to the introduction of Variable objects, which can carry
additional data, and Domains which act as a factory and lookup table
for these objects. The numbering of variables is implemented by
querying the domain. Domains can implement and cache domain-specifically
structured variables, like one variable for each 2-face of the n-cube.

Domains solve also the problem posed by the following expression

  Formula("a") & Formula("~a")

Do the two "a" variables refer to the same variable? How to tell?
Each formula takes its variables out of a specific domain and only
formulas with the same domain may be combined with a connective.
In the above code, the implicit default domain is Propcalc::DefaultDomain
which is an instance of Propcalc::Cache. The Cache implements a Domain
by allocating a new variable whenever a hitherto unknown variable is
asked for from the domain and caching that variable for future queries,
also reporting consistent variable numbers.

The concept of non-symbolic variables is removed and with it the
interface problems that arise with heterogeneous variable types.

Also add infix stringification support and add "introspection"
methods to the Ast::* classes which allow to determine a node's
type, associativity and precedence at runtime.

The C wrapper will remain non-functional until the v0.1 release goals
are achieved in the C++ part.