Wrap Variable in Ast::Var node
This is so that we can remove some shared_ptr's from the interface
of Domain, because in reality the Domain owns the variables and it
should just give out pointers for anyone asking.

Using make_unique bumps the C++ standard requirements to C++14...

Also remove the get_nr member from Domain for now. The future will
tell what we need in terms of variable ordering and how it should
be implemented.