Cache: keep insertion order of variables
The order of variables plays a big role in applications I have in mind.
(The domain-specific variables must follow a specific ordering so that
output of AllSAT solvers are compatible with other files we created in
the past.) This will be stressed later in the interface of Domain, and
the Cache implementation already adheres to it now.