Implement lazy CNF clauses generator for Formula
Currently not an iterator, same as Truthtable, but in the future it
should become one, I suppose. The only strategy implemented at the
moment is running through the formula's truth table and making clauses
from false assignments -- but using the observation that Ast::And nodes
at the root (and at other Ast::And nodes near the root, etc.) can be
skipped because those just implement concatenation of clauses of their
subtrees. [ For the applications I have in mind, this is already enough,
as I deal with small formulas that are axioms (so And-connected) and
which are just repeated under a group action, hence their clauses also
just repeat. ]

Later maybe clause rewriting (using DeMorgan and distributivity) might
be implemented and formulas might be JITted into assembly circuits when
they use few enough variables so that they all fit into a machine word,
in order to speed up truth table generation for these subformulas.