Add Tseitin transform
Like CNF, Tseitin is a Stream<Clause> lazily generating Clause objects
which implement the Tseitin transform of a formula. This is a formula
in CNF which is equisatisfiable but not equivalent to the original
formula, but whose size is always linearly bounded. This is useful
for SAT but not #SAT or AllSAT queries.