Negate CNF assignments to get clauses
This was an oversight because we don't have tests yet. The CNF stream
would return the false assignments to the formula as clauses, but their
negations should have been returned: the false assignments is what the
CNF wants to /forbid/.

Also adds an operator~ overload for Assignment to get the negated
assignment.