Redo Stream interface
The template Stream<T> remains typesafe in that clients know that the
return type of dereferencing it is a T *and* the Stream can give out
STL-compatible forward iterators now which are known to produce T's
as well.  Range-based for loops are so much nicer now.

This apparently also reduced the number of allocations in cpptest.