Cache::unpack also allocated variables
A Cache is a domain that tries to be universally usable and answer every
query with "yes, here it is" and remembers what it gave out.

This is how resolve works and unpack should follow that. Otherwise the
DIMACS reader will be unable to read a general formula for which the user
did not write a custom domain. That makes the library much less useful.
This commit makes unpack allocate as many variables as needed to give
back the requested variable number. It uses the variable number as the
variable name, in the absence of better information.