Implement Stream::Cached
The general Stream got a cache method which returns a Stream::Cached
object. This object is another Stream of the same type. On construction,
it iterates its input stream and puts all return values into a vector,
so that it can be iterated multiple times.

As a consequence, CNF and Tseitin had to change their API to return
Clause objects instead of const Clause&. This needs a few more
allocations but it seems cleaner as it allows me to also uncomment
some `using`s in the Stream::iterator interface... Not sure how to
do it better.