Fix Assignment confusion and bug
The confusion was in the interface: the operator[](VarNr) was not really
looking up the assignment to a given VarNr. The parameter was an index
into the Assignment's order member, not into a Domain as VarNr is.
This interface is completely removed because it had no real use. The one
occurrence in cpptest could be replaced by a neater range-based for.

The bug concerns the modifying overload of operator[](VarRef). This is
not actually modifying because it uses assign.at(), which throws if the
VarRef didn't exist yet. This is undesirable, as this overload should
create the mapping and return a reference to the value. (We must also
not forget to add the new VarRef to the order member.)