Prevent parser from accepting RPN
Just parsing with an operator stack and output deque does not allow us
to decide if the input is valid infix. It would accept some postfix as
well. This is now fixed by expecting terms and infixes in alternation.

This makes previously failing tests pass but we now test the reported
location in parser exceptions, too, which gives a few new TODOs.