Fix remaining parser tests
Be more strict about about expectation checks. The "X expected.*"
errors are now predominant in the user interface, as the tests show.
The original Shunting-yard errors do not really appear anymore,
but remain in the sources because they theoretically diagnose more
fine-grained errors. Fact is that the previous implementation of
Shunting-yard was not nearly robust enough.