CNF: evaluate only subformula
CNF skips AND nodes at the root and reduces the problem of generating
a CNF for a formula to subformulas not rooted at AND nodes (their CNFs
have to be concatenated to solve the original problem).

But the CNF algorithm evaluated the global formula for each assignment.
This is not only wasteful (as only the subformula has to be evaluated),
but could also crash if the Assignment created from variables found in
the subformula was queried for non-contained variables from other parts
of the AST of the whole forumla.