Un-shared_ptr Domain references
This doesn't really have an impact on memory usage or performance,
but it feels cleaner not to expose memory management through the API,
also with respect to FFI later. I expect that applications use either
lexically scoped temporary or statically allocated singleton domains
that live as long as the process. Also expose root and domain in
Formula and get rid of that get_domain() method.