Store Formula as a vector instead of tree
This commit changes the internal representation of Formula from a tree
of AST nodes to a compact vector of 32-bit AST data structures. Doing so
promised to speed things up and reduce memory usage, but it turne out
to be not so easy because some algorithms (in particular the Tseitin
transform) would require many "materializations" of subformulas for
intricate technical reasons.

Suffice it to say that I put this commit into a separate branch for now
and show the performance of tests before and after:

Before (the parent of this commit):

$ prove -e '/usr/bin/valgrind --error-exitcode=111 --exit-on-first-error=yes --leak-resolution=low --leak-check=full --errors-for-leak-kinds=all' /home/tab/src/SAT/libpropcalc/build/domain.t /home/tab/src/SAT/libpropcalc/build/formula.t /home/tab/src/SAT/libpropcalc/build/parser.t
/home/tab/src/SAT/libpropcalc/build/domain.t ...
==3961415==   total heap usage: 120 allocs, 120 frees, 82,302 bytes allocated
/home/tab/src/SAT/libpropcalc/build/formula.t ...
==3961425==   total heap usage: 2,134,066 allocs, 2,134,066 frees, 56,288,786 bytes allocated
/home/tab/src/SAT/libpropcalc/build/parser.t ...
==3961926==   total heap usage: 1,577 allocs, 1,577 frees, 326,548 bytes allocated
Files=3, Tests=25, 56 wallclock secs ( 0.12 usr  0.03 sys + 56.18 cusr  0.18 csys = 56.51 CPU)

After (this commit):

/home/tab/src/SAT/libpropcalc/build/domain.t ...
==3696145==   total heap usage: 119 allocs, 119 frees, 82,070 bytes
/home/tab/src/SAT/libpropcalc/build/formula.t ...
==3696154==   total heap usage: 3,115,760 allocs, 3,115,760 frees, 73,287,368 bytes allocated
/home/tab/src/SAT/libpropcalc/build/parser.t ...
==3696707==   total heap usage: 2,372 allocs, 2,372 frees, 339,194 bytes
Files=3, Tests=25, 80 wallclock secs ( 0.14 usr  0.02 sys + 79.26 cusr  0.25 csys = 79.67 CPU)