Add connectives as methods to Formula
The Formula class now has methods notf, andf, orf, thenf, eqvf and xorf
to build new formulas. The "f" suffix is pretty much arbitrary and just
protects against not, and, or and xor being recognized as keywords.

The overloaded operators have been cut down. operator>> is not used
anymore as it is too strange and operator== has been repurposed to
check structural equality of Formula objects.

Also enable the extra tests for CNF in formula.t.