Build caching into Stream<T>
Previously, every Stream<T> could defer to a caching version of itself
of type Stream<T>::Cached by calling the cache() method. However, this
would erase the type of derived streams (which are all our streams).

Instead, make caching a part of the Stream<T> template which can be
turned on by instance.