Specialize Stream<Clause> to Conjunctive
Conjunctive is an instantiated Stream<Clause> with additional methods.
It represents a lazy stream of clauses which should be ANDed, as in
CNF and Tseitin and when reading DIMACS files. Because it appears so
often, it is good to have a dedicated type for it that can be extended.
In particular, Conjunctive (and now also Clause) have an eval method.

This simplifies formula.t tests but it also speeds them up remarkably:

  Before
  ==============================================================================
  stream.t ... total heap usage: 26 allocs, 26 frees, 77,534 bytes allocated
  domain.t ... total heap usage: 119 allocs, 119 frees, 82,070 bytes allocated
  formula.t .. total heap usage: 2,179,168 allocs, 2,179,168 frees, 57,430,550 bytes allocated
  parser.t ... total heap usage: 1,575 allocs, 1,575 frees, 326,084 bytes allocated
  Files=4, Tests=28, 59 wallclock secs ( 0.14 usr  0.02 sys + 59.22 cusr  0.16 csys = 59.54 CPU)

  After
  ==============================================================================
  stream.t ... total heap usage: 26 allocs, 26 frees, 77,534 bytes allocated
  domain.t ... total heap usage: 119 allocs, 119 frees, 82,070 bytes allocated
  formula.t .. total heap usage: 1,208,882 allocs, 1,208,882 frees, 35,240,462 bytes allocated
  parser.t ... total heap usage: 1,575 allocs, 1,575 frees, 326,084 bytes allocated
  Files=4, Tests=28, 35 wallclock secs ( 0.13 usr  0.02 sys + 35.37 cusr  0.23 csys = 35.75 CPU)

The gain in runtime seems proportional to the loss in allocations in
the big test (CNF & Tseitin).